Split index.ts to index.ts/cli.ts
1. It's intend to use faker in express conveniently, such as:

``` javascript
const { applyFaker } = require('graphql-faker')
const express = require('express')
const app = express()
// your codes here
applyFaker(app, { fileArg: '/path/to/schema.graphql' })
// more codes
```

2. Use cli.js as bin
