+++ package.json
@@ -3,11 +3,11 @@
   "version": "1.5.0",
   "description": "Mock or extend your GraphQL API with faked data. No coding required",
   "main": "dist/index.js",
-  "bin": "dist/index.js",
+  "bin": "dist/cli.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon -e ts --exec 'ts-node src/index.ts'",
-    "debug": "ts-node --inspect --compilerOptions '{\"inlineSources\":true}' src/index.ts",
+    "start": "nodemon -e ts --exec 'ts-node src/cli.ts'",
+    "debug": "ts-node --inspect --compilerOptions '{\"inlineSources\":true}' src/cli.ts",
     "build:editor": "cd src/editor && npm install && npm run build && cd -",
     "build:typescript": "tsc",
     "copy:graphql": "cp src/*.graphql dist/",
